feat(demo): First work on plain ts demo
