--- demo/app/examples/NSChart.xml
@@ -1,20 +1,0 @@
-<Page xmlns="http://schemas.nativescript.org/tns.xsd" xmlns:chart="@nativescript-community/ui-chart">
-    <ActionBar title="NS Chart">
-        <NavigationButton text="Back" android.systemIcon="ic_menu_back" tap="onNavigationButtonTap"/>
-        <Button text="Redraw" tap="redraw" />
-    </ActionBar>
-    <ScrollView>
-        <StackLayout>
-            <Label text="Radar Chart" fontSize="20"/>
-            <chart:RadarChart id="radar-chart" backgroundColor="lightgray" width="300" height="350" loaded="onRadarChartLoaded"/>
-            <Label text="Pie Chart" fontSize="20"/>
-            <chart:PieChart id="pie-chart" backgroundColor="lightgray" width="300" height="350" loaded="onPieChartLoaded"/>
-            <Label text="Line Chart" fontSize="20"/>
-            <chart:LineChart id="line-chart" backgroundColor="lightgray" width="300" height="350" loaded="onLineChartLoaded"/>
-            <Label text="Bar Chart" fontSize="20"/>
-            <chart:BarChart id="bar-chart" backgroundColor="lightgray" width="300" height="350" loaded="onBarChartLoaded"/>
-            <Label text="Horizontal Bar Chart" fontSize="20"/>
-            <chart:HorizontalBarChart id="horizontal-bar-chart" backgroundColor="lightgray" width="300" height="350" loaded="onBarChartLoaded"/>
-        </StackLayout>
-    </ScrollView>
-</Page>
