--- demo/app/examples/NSUIChart.xml
@@ -1,24 +1,0 @@
-<Page xmlns="http://schemas.nativescript.org/tns.xsd" xmlns:chart="nativescript-ui-chart" loaded="onLoaded">
-    <ActionBar title="NS UI Chart">
-        <NavigationButton text="Back" android.systemIcon="ic_menu_back" tap="onNavigationButtonTap"/>
-    </ActionBar>
-    <ScrollView>
-        <StackLayout>
-            <chart:RadCartesianChart width="300" height="400">
-                <chart:RadCartesianChart.series>
-                    <chart:LineSeries seriesName="Line" items="{{ data }}" legendTitle="value" categoryProperty="index" valueProperty="value">
-                        <chart:LineSeries.horizontalAxis>
-                            <chart:CategoricalAxis allowPan="true" allowZoom="true" />
-                        </chart:LineSeries.horizontalAxis>
-                        <chart:LineSeries.verticalAxis>
-                            <chart:LinearAxis allowPan="true" allowZoom="true" />
-                        </chart:LineSeries.verticalAxis>
-                    </chart:LineSeries>
-                </chart:RadCartesianChart.series>
-                <chart:RadCartesianChart.legend>
-                    <chart:RadLegendView position="Bottom" width="150"/>
-                </chart:RadCartesianChart.legend>
-            </chart:RadCartesianChart>
-        </StackLayout>
-    </ScrollView>
-</Page>
