+++ demo/package.json
@@ -10,7 +10,6 @@
         "@nativescript/theme": "~2.3.3",
         "@nativescript-community/ui-chart": "../plugin",
         "nativescript-perms": "^2.0.11",
-        "nativescript-ui-chart": "8.0.2",
         "typescript": "4.0.3"
     },
     "devDependencies": {
